+++ server/norace_test.go
@@ -1,0 +1,85 @@
+// Copyright 2018 The NATS Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// +build !race
+
+package server
+
+import (
+	"fmt"
+	"math/rand"
+	"sync/atomic"
+	"testing"
+	"time"
+
+	"github.com/nats-io/go-nats"
+)
+
+// IMPORTANT: Tests in this file are not executed when running with the -race flag.
+
+func TestAvoidSlowConsumerBigMessages(t *testing.T) {
+	opts := DefaultOptions() // Use defaults to make sure they avoid pending slow consumer.
+	s := RunServer(opts)
+	defer s.Shutdown()
+
+	nc1, err := nats.Connect(fmt.Sprintf("nats://%s:%d", opts.Host, opts.Port))
+	if err != nil {
+		t.Fatalf("Error on connect: %v", err)
+	}
+	defer nc1.Close()
+
+	nc2, err := nats.Connect(fmt.Sprintf("nats://%s:%d", opts.Host, opts.Port))
+	if err != nil {
+		t.Fatalf("Error on connect: %v", err)
+	}
+	defer nc2.Close()
+
+	data := make([]byte, 1024*1024) // 1MB payload
+	rand.Read(data)
+
+	expected := int32(1000)
+	received := int32(0)
+
+	done := make(chan bool)
+
+	// Create Subscription.
+	nc1.Subscribe("slow.consumer", func(m *nats.Msg) {
+		// Just eat it so that we are not measuring
+		// code time, just delivery.
+		atomic.AddInt32(&received, 1)
+		if received >= expected {
+			done <- true
+		}
+	})
+
+	// Create Error handler
+	nc1.SetErrorHandler(func(c *nats.Conn, s *nats.Subscription, err error) {
+		t.Fatalf("Received an error on the subscription's connection: %v\n", err)
+	})
+
+	for i := int32(0); i < expected; i++ {
+		nc2.Publish("slow.consumer", data)
+	}
+	nc2.Flush()
+
+	select {
+	case <-done:
+		return
+	case <-time.After(10 * time.Second):
+		r := atomic.LoadInt32(&received)
+		if s.NumSlowConsumers() > 0 {
+			t.Fatalf("Did not receive all large messages due to slow consumer status: %d of %d", r, expected)
+		}
+		t.Fatalf("Failed to receive all large messages: %d of %d\n", r, expected)
+	}
+}
