+++ test/cluster_test.go
@@ -488,43 +488,3 @@ func TestAutoUnsubscribePropagationOnClientDisconnect(t *testing.T) {
 		t.Fatalf("%v", err)
 	}
 }
-
-func TestRouteFormTimeWithHighSubscriptions(t *testing.T) {
-	srvA, optsA := RunServerWithConfig("./configs/srv_a.conf")
-	defer srvA.Shutdown()
-
-	clientA := createClientConn(t, optsA.Host, optsA.Port)
-	defer clientA.Close()
-
-	sendA, expectA := setupConn(t, clientA)
-
-	// Now add lots of subscriptions. These will need to be forwarded
-	// to new routes when they are added.
-	subsTotal := 100000
-	for i := 0; i < subsTotal; i++ {
-		subject := fmt.Sprintf("FOO.BAR.BAZ.%d", i)
-		sendA(fmt.Sprintf("SUB %s %d\r\n", subject, i))
-	}
-	sendA("PING\r\n")
-	expectA(pongRe)
-
-	srvB, _ := RunServerWithConfig("./configs/srv_b.conf")
-	defer srvB.Shutdown()
-
-	checkClusterFormed(t, srvA, srvB)
-
-	// Now wait for all subscriptions to be processed.
-	if err := checkExpectedSubs(subsTotal, srvB); err != nil {
-		// Make sure we are not a slow consumer
-		// Check for slow consumer status
-		if srvA.NumSlowConsumers() > 0 {
-			t.Fatal("Did not receive all subscriptions due to slow consumer")
-		} else {
-			t.Fatalf("%v", err)
-		}
-	}
-	// Just double check the slow consumer status.
-	if srvA.NumSlowConsumers() > 0 {
-		t.Fatalf("Received a slow consumer notification: %d", srvA.NumSlowConsumers())
-	}
-}
